pyproject.toml
[project]
name = "secret-santa"
version = "0.1.1"
description = "Self-hosted Secret Santa gift exchange organizer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlmodel==0.0.14",
    "aiosqlite==0.19.0",
    "authlib==1.2.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "jinja2==3.1.2",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "python-telegram-bot==20.7",
    "itsdangerous>=2.0.0",
    "greenlet>=3.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
# Note: dev-dependencies is deprecated, but kept for compatibility
# New projects should use dependency-groups instead
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# Modern way (for future uv versions)
[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

